backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
firebase-admin==6.5.0
python-dotenv==1.0.1
requests==2.31.0
google-genai==0.3.0
